Use another return type from api
