+++ src/api/learningpathApi.ts
@@ -30,7 +30,7 @@ export async function fetchLearningpaths(
   });
 }
 
-export async function fetchMyLearningpaths(context: Context): Promise<Array<ILearningPathSummaryV2>> {
+export async function fetchMyLearningpaths(context: Context): Promise<Array<ILearningPathV2>> {
   const response = await fetch(
     `/learningpath-api/v2/learningpaths/mine?language=${context.language}&fallback=true`,
     context,
@@ -38,25 +38,10 @@ export async function fetchMyLearningpaths(context: Context): Promise<Array<ILea
   return await resolveJson(response);
 }
 
-export async function fetchLearningpath(id: string, context: Context): Promise<GQLLearningpath> {
+export async function fetchLearningpath(id: string, context: Context): Promise<ILearningPathV2> {
   const response = await fetch(
     `/learningpath-api/v2/learningpaths/${id}?language=${context.language}&fallback=true`,
     context,
   );
-  const learningpath: ILearningPathV2 = await resolveJson(response);
-  const learningsteps = learningpath.learningsteps.map((step) => ({
-    ...step,
-    title: step.title.title,
-    description: step.description?.description,
-  }));
-
-  return {
-    ...learningpath,
-    title: learningpath.title.title,
-    description: learningpath.description.description,
-    lastUpdated: learningpath.lastUpdated,
-    coverphoto: learningpath.coverPhoto,
-    tags: learningpath.tags.tags || [],
-    learningsteps,
-  };
+  return await resolveJson(response);
 }
